Bug fix for MSVC Error	1	error C2440: '<function-style-cast>' : cannot convert from 'double' to 'Depth'	search.cpp	387	1

+++ src/search.cpp
@@ -384,11 +384,11 @@ void Thread::search(bool isMainThread) {
           ++depth;
           for (Thread* th : Threads)
               if (th != this && th->depth <= depth)
-                  th->depth = depth + Depth(3 * log(1 + th->idx));
+                  th->depth = depth + Depth(int(3 * log(1 + th->idx)));
       }
       else
           // This can cause a thread to search with the same depth for many iterations
-          depth = Threads.main()->depth + Depth(3 * log(1 + this->idx));
+          depth = Threads.main()->depth + Depth(int(3 * log(1 + this->idx)));
 
       if (depth >= DEPTH_MAX || Signals.stop || (Limits.depth && depth > Limits.depth))
           break;
